+++ tests/data.rs
@@ -40,10 +40,8 @@ impl<'a, 'b> SimpleState<'a, 'b> for TestState {
 
 fn run_world(num_entities: usize) -> amethyst::Result<()> {
     let editor_sync_bundle = SyncEditorBundle::new();
-    let game_data = GameDataBuilder::default()
-        .with_bundle(editor_sync_bundle)?;
-    let mut game = Application::build(".", TestState::new(num_entities))?
-        .build(game_data)?;
+    let game_data = GameDataBuilder::default().with_bundle(editor_sync_bundle)?;
+    let mut game = Application::build(".", TestState::new(num_entities))?.build(game_data)?;
     game.run();
 
     Ok(())